# openapi.yaml
openapi: 3.0.3
info:
  title: API ИИ-гида музея Фаберже
  version: 0.1.0-mvp
  description: |
    Backend-API мобильного web-приложения (PWA) «ИИ-гид музея Фаберже».

    Посетитель сканирует **QR-код** на входе экспозиции, попадает в приложение
    без регистрации и установки, перемещается по **интерактивной карте**
    (зал → витрина → экспонат), а главное — может **сфотографировать экспонат**,
    получить о нём рассказ от ИИ-гида и **прослушать** его озвучку.

    Этот документ описывает контракт backend-сервиса (FastAPI на Yandex Cloud),
    к которому обращается фронтенд. Спецификация написана по дорожной карте MVP
    (25 мая – 25 июня) и служит единым контрактом между Frontend, Backend и ML.

    ## Ключевой пользовательский сценарий (бесшовное распознавание)

    ```
    Клик «Распознать экспонат»
      → камера (WebRTC)
      → снимок
      → POST /recognition (фото → YOLO → label_slug)
      → POST /guide/story (label_slug → raw_history из БД → YandexGPT → рассказ)
      → показ рассказа + 3–4 всплывающих вопроса
      → POST /speech («Прослушать», SpeechKit → аудио)
      → блок «Другие экспонаты зала» (GET /exhibits/{id}/related)
    ```

    ## Доменная модель

    Иерархия каталога повторяет физическую структуру музея:

    | Уровень   | Сущность   | Таблица БД   | Ключевые поля                                   |
    |-----------|------------|--------------|-------------------------------------------------|
...

...
    (через публичный API не отдаётся).

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security: []
paths:
...
  /exhibits:
    get:
      tags: [Экспонаты]
      summary: Список / фильтрация экспонатов
      description: >-
        Постраничный список экспонатов с фильтрами по залу, витрине,
        `label_slug` и подстроке названия. Универсальная выборка для каталога.
      operationId: listExhibits
      security: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: hall_id
          in: query
          description: Фильтр по идентификатору зала.
          required: false
          schema:
            type: integer
            minimum: 1
          example: 5
        - name: showcase_id
          in: query
          description: Фильтр по идентификатору витрины.
          required: false
          schema:
            type: integer
            minimum: 1
        - name: label_slug
          in: query
          description: Фильтр по точному `label_slug` (то, что возвращает YOLO).
          required: false
          schema:
            type: string
          example: faberge_egg_coronation
        - name: q
          in: query
          description: Поиск по подстроке названия экспоната.
          required: false
          schema:
            type: string
          example: яйцо
        - $ref: '#/components/parameters/IsTemporary'
      responses:
        '200':
          description: Список экспонатов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitListResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /exhibits/{exhibit_id}:
    get:
      tags: [Экспонаты]
      summary: Карточка экспоната
      description: >-
        Полные данные карточки. Поле `raw_history` в публичный ответ не включается.


        Состав и порядок полей музей задал дословно (баг-репорт 31.08.2026,
        п. I-2): название предмета, изображение, расположение (название и номер
        зала, номер витрины), год создания, фирма и мастер, материалы, техники,
        описание. Ответ идёт ровно в этом порядке.


        Что важно фронту:

        • расположение приходит ГОТОВОЙ строкой в `location.text`
        («Зал 4 «Синяя гостиная», витрина 5») — собирать его из `hall`/`showcase`
        не нужно: ту же фразу печатает ИИ-гид, и две склейки разойдутся;

        • `maker.text` всегда содержит полную строку «фирма и мастер» как её ведёт
        музей, `maker.firm`/`maker.master` — её разобранные части (могут быть
        `null`, тогда показывайте `maker.text`);

        • набор ключей ПОСТОЯНЕН: пустое значение приходит как `null`, ключ не
        исчезает, `location` и `maker` — всегда объекты. Проверки «а есть ли поле»
        писать не нужно.
      operationId: getExhibit
      security: []
      parameters:
        - $ref: '#/components/parameters/ExhibitId'
      responses:
        '200':
          description: Карточка экспоната.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exhibit'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /search:
    get:
      tags: [Поиск]
      summary: Поиск по залам и экспонатам
      description: >-
        Поисковая строка интерактивной карты: ищет совпадения в названиях
        залов и экспонатов. Возвращает две раздельные коллекции результатов.
      operationId: search
      security: []
      parameters:
        - name: q
          in: query
          description: Поисковый запрос (название зала или экспоната).
          required: true
          schema:
            type: string
            minLength: 1
          example: коронационное
        - name: limit
          in: query
          description: Максимум результатов в каждой коллекции.
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: Результаты поиска.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  # ─────────────────────────── Распознавание ─────────────────────────────
  /recognition:
    post:
      tags: [Распознавание]
      summary: Распознать экспонат по фото
      description: |
        Принимает фотографию экспоната, отправляет её в развёрнутую на
        Yandex Cloud модель **YOLO** и возвращает распознанный `label_slug`
        вместе с найденной в БД карточкой экспоната.

        Сразу после снимка фронтенд бесшовно переходит в окно чата с
        состоянием загрузки и вызывает этот эндпоинт. При успешном
        распознавании далее обычно вызывается `POST /guide/story`.

        Если уверенность ниже порога — `recognized = false`; в `candidates`
        могут вернуться наиболее вероятные классы для уточнения.

        **Лимит на файл — 2,5 МБ** (`MAX_UPLOAD_MB`), превышение даёт 413 с
        текстом «Фото слишком большое — X МБ при допустимых 2,5 МБ». Он
        намеренно ниже платформенной границы Yandex API Gateway — **3,5 МиБ
        (3 670 016 байт) на весь запрос**, столько же на payload синхронного
        вызова Cloud Functions, куда тело уезжает в base64 (+33 %). Это не наш
        выбор и кодом не поднимается — только конфигурацией API Gateway.

        Почему граница важна для фронта: гейтвей режет запрос ДО приложения и
        отвечает своим 413 без CORS-заголовков, поэтому в браузере это не «фото
        слишком большое», а «Failed to fetch». Пока наш лимит ниже
        платформенного, 413 отдаёт сам сервис — с CORS и внятным текстом
        (п. 10 баг-репорта 06.08.2026). Кадр с камеры телефона (3–6 МБ) нужно
        сжимать перед отправкой; кадр из canvas на экране распознавания
        (~0,5 МБ) проходит как есть.
      operationId: recognizeExhibit
      security: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Фото экспоната (JPEG/PNG/WebP, до 2,5 МБ).
                hall_id:
                  type: integer
                  minimum: 1
                  description: >-
                    Необязательная подсказка о текущем зале (по карте) —
                    повышает точность и сужает поиск.
                top_k:
                  type: integer
                  minimum: 1
                  maximum: 10
                  default: 3
                  description: Сколько кандидатов вернуть при неуверенном распознавании.
            encoding:
              file:
                contentType: image/jpeg, image/png, image/webp
      responses:
        '200':
          description: Результат распознавания (успех или «не распознано»).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecognitionResponse'
              examples:
                recognized:
                  summary: Экспонат распознан
                  value:
                    recognized: true
                    label_slug: faberge_egg_coronation
                    confidence: 0.94
                    request_id: 7b1c0e2a-9f3d-4a2b-8c10-2a4d6e8f0a11
                    processing_ms: 312
                    exhibit:
                      id: 101
                      label_slug: faberge_egg_coronation
                      name: Яйцо «Коронационное»
                      image_url: https://cdn.example.cloud/exhibits/coronation/main.jpg
                      location:
                        hall_id: 4
                        hall_number: 4
                        hall_name: Синяя гостиная
                        showcase_id: 12
                        showcase_number: 5
                        showcase_name: Пасхальные подарки
                        text: Зал 4 «Синяя гостиная», витрина 5
                        text_in: в зале 4 «Синяя гостиная», витрина 5
                      hall:
                        id: 4
                        hall_number: 4
                        name: Синяя гостиная
                      showcase:
                        id: 12
                        showcase_number: 5
                        name: Пасхальные подарки
                      year_created: '1897'
                      origin_place: Санкт-Петербург
                      maker:
                        text: Фирма К. Фаберже, мастер М. Перхин
                        firm: Фирма К. Фаберже
                        master: мастер М. Перхин
                      master_name: Фирма К. Фаберже, мастер М. Перхин
                      material: Золото, гильоше-эмаль, бриллианты, рубины
                      techniques: гравировка, канфарение, эмаль по гильошированному фону
                      short_description: >-
                        Пасхальное яйцо-сюрприз, подаренное Николаем II
                        императрице Александре Фёдоровне в память о коронации.
                      model_3d_url: https://koinovo.ru/fabergemuseum/coronation
                    candidates: []
                not_recognized:
                  summary: Не распознано — есть кандидаты
                  value:
                    recognized: false
                    label_slug: null
                    confidence: 0.41
                    request_id: 0c5b2a18-3e7f-4d9a-9b21-7c3e1f5a2b00
                    processing_ms: 287
                    exhibit: null
                    candidates:
                      - label_slug: faberge_egg_winter
                        name: Яйцо «Зимнее»
                        confidence: 0.41
                      - label_slug: faberge_egg_lilies
                        name: Яйцо «Ландыши»
                        confidence: 0.33
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '502':
          $ref: '#/components/responses/UpstreamError'

  # ───────────────────────────── ИИ-гид ──────────────────────────────────
...
  /guide/chat:
    post:
      tags: [ИИ-гид]
      summary: Диалог с ИИ-гидом
      description: |
        Многоходовой диалог с гидом. Контекст задаётся экспонатом
        (`exhibit_id`/`label_slug`) либо залом (`hall_id` — для кнопки
        «Спросить AI-гида» на уровне зала). Часто `message` — это один из
        ранее предложенных `suggested_questions`.

        Состояние диалога хранится по `session_id`: если он не передан, сервер
        создаёт новый и возвращает его. Альтернативно можно передавать
        историю сообщений в `history`. Ответ снова содержит набор
        контекстных вопросов-подсказок.
      operationId: chatWithGuide
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            examples:
              followUp:
                summary: Уточняющий вопрос по экспонату
                value:
                  session_id: 2f6a9b3c-1d4e-4f8a-9c2b-5e7a1d0c3b66
                  context:
                    exhibit_id: 101
                  message: Кому подарили это яйцо?
      responses:
        '200':
          description: Ответ гида и новые подсказки.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
              examples:
                answer:
                  summary: Ответ на уточняющий вопрос
                  value:
                    session_id: 2f6a9b3c-1d4e-4f8a-9c2b-5e7a1d0c3b66
                    answer: >-
                      Это яйцо Николай II подарил своей супруге, императрице
                      Александре Фёдоровне, на Пасху 1897 года…
                    suggested_questions:
                      - А что подарили его матери?
                      - Где оно хранилось после революции?
                      - Сколько таких яиц всего создал Фаберже?
                    context:
                      exhibit_id: 101
                      label_slug: faberge_egg_coronation
                    referenced_exhibits:
                      - id: 101
                        name: Яйцо «Коронационное»
                        exhibit_number: '12'
                        thumbnail_url: https://cdn.example.cloud/exhibits/coronation/thumb.jpg
                        hall_number: 4
                        showcase_number: 5
                        mentioned_in: context
                    referenced_halls: []
                    location: null
                noMentions:
                  summary: Ответ ни на один предмет каталога не ссылается — блок «Упомянуто в ответе» не показываем
                  value:
                    session_id: 2f6a9b3c-1d4e-4f8a-9c2b-5e7a1d0c3b66
                    answer: >-
                      Музей Фаберже открылся в Шуваловском дворце в 2013 году.
                      Основу собрания составила коллекция Виктора Вексельберга…
                    # Общий чат без контекста экспоната: подсказок здесь раньше
                    # не было вовсе (п. II-7). Теперь отдаётся детерминированный
                    # набор про музей — без обращения к модели.
                    suggested_questions:
                      - Какие залы есть в музее?
                      - Что обязательно посмотреть в коллекции?
                      - Чем знаменит Карл Фаберже?
                    context: null
                    referenced_exhibits: []
                    referenced_halls: []
                    location: null
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '502':
          $ref: '#/components/responses/UpstreamError'

  # ──────────────────────────── Озвучивание ──────────────────────────────
...
  /admin/showcases:
    post:
      tags: [Администрирование]
      summary: '[Вне MVP] Создать витрину'
      operationId: createShowcase
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShowcaseCreate'
      responses:
        '201':
          description: Витрина создана.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowcaseDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /admin/showcases/{showcase_id}:
    patch:
      tags: [Администрирование]
      summary: '[Вне MVP] Частично обновить витрину'
      description: >-
        Частичное обновление витрины (E14): любой поднабор
        `{hall_id, showcase_number, name}`. Перенос в другой зал и смена номера
        учитывают уникальность `(hall_id, showcase_number)` — при конфликте `409`.
      operationId: updateShowcase
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ShowcaseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShowcasePatch'
      responses:
        '200':
          description: Витрина обновлена.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowcaseDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    delete:
      tags: [Администрирование]
      summary: '[Вне MVP] Удалить витрину'
      description: >-
        Удаляет витрину. По умолчанию, если в витрине есть экспонаты, возвращает
        `409` — сначала опустошите витрину или передайте `?force=true`. При
        `force=true` каскадно удаляются экспонаты и их фото (включая объекты в
        Object Storage).
      operationId: deleteShowcase
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ShowcaseId'
        - name: force
          in: query
          required: false
          description: Каскадно удалить экспонаты и их медиа.
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: Витрина удалена.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/exhibits:
    post:
      tags: [Администрирование]
      summary: '[Вне MVP] Создать экспонат'
      description: >-
        Создание экспоната, включая внутреннее поле `raw_history`
        (факты для генерации рассказа LLM).
      operationId: createExhibit
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExhibitCreate'
            examples:
              winterEgg:
                summary: Зимнее яйцо с raw_history
                value:
                  showcase_id: 12
                  label_slug: faberge_egg_winter
                  name: Яйцо «Зимнее»
                  year_created: '1913'
                  master_name: Альма Пиль
                  material: Горный хрусталь, платина, бриллианты
                  techniques: резьба по камню, гравировка, монтировка
                  short_description: >-
                    Пасхальное яйцо в виде ледяного кристалла с сюрпризом-корзиной
                    подснежников.
                  raw_history: >-
                    Заказчик — Николай II. Подарок императрице Марии Фёдоровне.
                    Дизайн — Альма Пиль. Использованы горный хрусталь и бриллианты,
                    имитирующие иней. Сюрприз — корзина подснежников.
      responses:
        '201':
          description: Экспонат создан.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitAdmin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /telemetry/events:
    post:
      tags: [Телеметрия]
      summary: Отправить события
      description: >-
        Пакетный приём анонимных пользовательских событий — источник данных для
        административной аналитики. С 21.07.2026 — обязательная часть контракта:
        без событий GET /admin/analytics/engagement и /analytics/routes пусты (B11).


        Контракт с 03.08.2026:

        * не более 50 событий в батче, иначе 422;

        * событие с типом вне словаря НЕ роняет батч — оно отбрасывается,
        остальные записываются, число отброшенных возвращается в `rejected`;

        * устаревший `audio_play` нормализуется в канонический `tts_play`;

        * в `props` сохраняются только ключи из контракта события, остальные
        (в т.ч. user_agent, ip, url, geo) отбрасываются; `props.text` длиннее
        500 символов обрезается.


        Приватность: IP и User-Agent запроса не читаются и не сохраняются —
        ни в events, ни в логи (docs/analytics-privacy.md).
      operationId: ingestEvents
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventBatch'
      responses:
        '202':
          description: События приняты в обработку.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventIngestResult'
              example:
                accepted: 9
                rejected: 1
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Статический Bearer-токен администратора (получается через
        `POST /admin/login`). Слать в заголовке `Authorization: Bearer <token>`
        ко всем эндпоинтам `/admin/**` (кроме `/admin/login`).
  parameters:
    Limit:
      name: limit
      in: query
      description: Максимум элементов на странице.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      description: Смещение от начала выборки (постраничность).
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    IsTemporary:
      name: is_temporary
      in: query
      required: false
      description: >-
        Фильтр по типу экспозиции: `true` — временные (сменные) выставки,
        `false` — основная (постоянная) экспозиция. Без параметра возвращаются
        и те и другие. Для экспонатов признак берётся от зала (через витрину).
      schema:
        type: boolean
      example: true
...
    ExhibitId:
      name: exhibit_id
      in: path
      required: true
      description: Идентификатор экспоната.
      schema:
        type: integer
        minimum: 1
      example: 101
...
  responses:
    NotFound:
      description: Ресурс не найден.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Экспонат не найден.
    BadRequest:
      description: Некорректный запрос.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Не удалось прочитать изображение.
...
    UnsupportedMediaType:
      description: Неподдерживаемый формат файла.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Поддерживаются только JPEG, PNG и WebP.
    UnprocessableEntity:
      description: Ошибка валидации запроса (формат FastAPI).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTPValidationError'
...
  schemas:
...
    ExhibitListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ExhibitSummary'
        total:
          type: integer
          example: 24
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0

    # ── Поиск ──
    SearchResponse:
      type: object
      required: [query, halls, exhibits]
      properties:
        query:
          type: string
          example: коронационное
        halls:
          type: array
          items:
            $ref: '#/components/schemas/Hall'
        exhibits:
          type: array
          items:
            $ref: '#/components/schemas/ExhibitSummary'
        total:
          type: integer
          description: Суммарное число найденных объектов.
          example: 1

    # ── Распознавание ──
...
    RecognitionResponse:
      type: object
      required: [recognized]
      properties:
        recognized:
          type: boolean
          description: Распознан ли экспонат с уверенностью выше порога.
          example: true
        label_slug:
          type: string
          nullable: true
          example: faberge_egg_coronation
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          nullable: true
          example: 0.94
        exhibit:
          allOf:
            - $ref: '#/components/schemas/Exhibit'
          nullable: true
          description: Найденная карточка (если распознано), иначе `null`.
        candidates:
          type: array
          description: Вероятные классы при неуверенном распознавании.
          items:
            $ref: '#/components/schemas/RecognitionCandidate'
        request_id:
          type: string
          format: uuid
          description: Идентификатор запроса для трассировки.
          example: 7b1c0e2a-9f3d-4a2b-8c10-2a4d6e8f0a11
        processing_ms:
          type: integer
          nullable: true
          example: 312

    # ── ИИ-гид ──
...